+++ NEWS.md
@@ -1,3 +1,10 @@
+# AzureGraph 1.2.2
+
+- Add support for batch requests:
+  - Each individual request is stored in an object of R6 class `graph_request`.
+  - Add `call_batch_endpoint()` function and `ms_graph$call_batch_endpoint()` method for calling the batch endpoint with a list of requests.
+- Handle throttling (HTTP 429 errors) gracefully.
+
 # AzureGraph 1.2.1
 
 - Allow setting an optional limit to the number of objects returned by the private `ms_object$get_paged_list()` method.